Changed Navbar Structure
